refactor: move pipeline implementations to own module
